refactor(protocol): migrate tests to external tests/ directory
- Add test-support feature with test helpers
- Move all tests to crates/protocol/tests/protocol_transport.rs
- Remove inline #[cfg(test)] module from lib.rs
